refactor logger crate
Refactors the logger crate prior to review
