+++ src/rust/logger/src/backend/nop.rs
@@ -1,0 +1,15 @@
+// Copyright 2021 Twitter, Inc.
+// Licensed under the Apache License, Version 2.0
+// http://www.apache.org/licenses/LICENSE-2.0
+
+use crate::*;
+
+pub struct NopLogSender {}
+
+impl Log for NopLogSender {
+    fn enabled(&self, _: &log::Metadata<'_>) -> bool {
+        false
+    }
+    fn log(&self, _: &log::Record<'_>) {}
+    fn flush(&self) {}
+}
