+++ src/rust/logger/src/macros.rs
@@ -1,0 +1,6 @@
+#[macro_export]
+macro_rules! klog {
+    ($($arg:tt)*) => (
+        error!(target: "klog", $($arg)*);
+    )
+}
